秒杀结果返回  ：   redis 判断是否秒杀成功

+++ aircraft-service/src/main/java/com/travel/function/dao/OrderInfoDao.java
@@ -8,7 +8,7 @@ public interface OrderInfoDao {
 
     int insert(OrderInfo record);
 
-    int insertSelective(OrderInfo record);
+    Long insertSelective(OrderInfo record);
 
     OrderInfo selectByPrimaryKey(Long id);
